fix superstruct const accessed before const init caused by circular deps
decided to live with circular deps, since this project is kinda functional style
refactor a bit while fixing
