+++ src/ipfs.ts
@@ -1,7 +1,7 @@
 (window as any).global = window; // for ipfs-core to work
 import { create as createIPFS } from 'ipfs-core';
 
-import { Game } from './lib/game';
+import type { Game } from './lib/game';
 
 import { IPFS_OPTIONS } from './env';
 import { DBG_MODE } from './lib/dbg';
